+++ hasher-matcher-actioner/hmalib/lambdas/api/matches.py
@@ -18,7 +18,7 @@ from threatexchange.descriptor import ThreatDescriptor
 from threatexchange.signal_type.md5 import VideoMD5Signal
 from threatexchange.signal_type.pdq import PdqSignal
 
-from hmalib.models import MatchRecord
+from hmalib.common.models.pipeline import MatchRecord
 from hmalib.common.models.signal import PDQSignalMetadata, PendingOpinionChange
 from hmalib.common.logging import get_logger
 from hmalib.common.messages.match import BankedSignal
