+++ hasher-matcher-actioner/hmalib/lambdas/api/submit.py
@@ -26,7 +26,7 @@ from hmalib.common.content_sources import S3BucketContentSource
 from hmalib.common.models.content import ContentObject, ContentRefType
 from hmalib.common.logging import get_logger
 from hmalib.common.messages.submit import URLSubmissionMessage
-from hmalib.models import PipelineHashRecord
+from hmalib.common.models.pipeline import PipelineHashRecord
 
 logger = get_logger(__name__)
