+++ hasher-matcher-actioner/hmalib/lambdas/hashing.py
@@ -19,7 +19,7 @@ from hmalib.common.logging import get_logger
 from hmalib import metrics
 from hmalib.common.messages.submit import URLSubmissionMessage
 from hmalib.hashing.unified_hasher import UnifiedHasher
-from hmalib.models import PipelineHashRecord
+from hmalib.common.models.pipeline import PipelineHashRecord
 from hmalib.common.content_sources import URLContentSource
 
 logger = get_logger(__name__)
