+++ hasher-matcher-actioner/hmalib/lambdas/matcher.py
@@ -1,9 +1,6 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 
 import functools
-from hmalib.common.config import HMAConfig
-from hmalib.hashing.unified_hasher import ContentSignal
-from hmalib.models import PipelineHashRecord
 import json
 import os
 import boto3
@@ -16,6 +13,9 @@ from threatexchange.signal_type.pdq import PdqSignal
 from hmalib import metrics
 from hmalib.common.logging import get_logger
 from hmalib.matchers.matchers_base import Matcher
+from hmalib.common.config import HMAConfig
+from hmalib.hashing.unified_hasher import ContentSignal
+from hmalib.common.models.pipeline import PipelineHashRecord
 
 INDEXES_BUCKET_NAME = os.environ["INDEXES_BUCKET_NAME"]
 DYNAMODB_TABLE = os.environ["DYNAMODB_TABLE"]
